refactor: MVC 패키지 분리
