+++ src/main/java/racingcar/Main.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import racingcar.controller.GameManager;
+
 public class Main {
     public static void main(String[] args) {
         GameManager.playGame();
