+++ src/main/java/racingcar/domain/CarRecord.java
@@ -1,7 +1,4 @@
-package racingcar;
-
-import java.util.ArrayList;
-import java.util.List;
+package racingcar.domain;
 
 public class CarRecord {
     private final String name;
