+++ src/main/java/racingcar/domain/RacingManager.java
@@ -1,6 +1,5 @@
-package racingcar;
+package racingcar.domain;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class RacingManager {
