+++ src/main/java/racingcar/domain/RacingRoundRecord.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import java.util.ArrayList;
 import java.util.List;
