+++ src/main/java/racingcar/domain/movable/Movable.java
@@ -1,4 +1,4 @@
-package racingcar.movable;
+package racingcar.domain.movable;
 
 public interface Movable {
     boolean canMove();
