+++ src/main/java/racingcar/domain/number_generator/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.number_generator;
+package racingcar.domain.number_generator;
 
 public interface NumberGenerator {
     int generate();
