+++ src/main/java/racingcar/domain/number_generator/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.number_generator;
+package racingcar.domain.number_generator;
 
 import java.util.Random;
